storage/estimate-usage-details-application-cache.https.tentative.html

<html manifest="resources/appcache.manifest">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const initPromise = new Promise(resolve => {
  applicationCache.addEventListener('cached', resolve, false);
  applicationCache.addEventListener('noupdate', resolve, false);
});

promise_test(async t => {
  const estimate = await navigator.storage.estimate();
  assert_true(typeof estimate === 'object');
  assert_true('usage' in estimate);
  assert_equals(typeof estimate.usage, 'number');
  assert_true('quota' in estimate);
  assert_equals(typeof estimate.quota, 'number');
}, 'estimate() resolves to dictionary with members');

promise_test(async t => {
  let estimate = await navigator.storage.estimate();

  const usageBeforeCreate = estimate.usageDetails.applicationCache || 0;

  await initPromise;

  estimate = await navigator.storage.estimate();
  const usageAfterCreate = estimate.usageDetails.applicationCache;

  assert_greater_than(
    usageAfterCreate, usageBeforeCreate);
}, 'estimate() shows usage increase after app is cached');
</script>
</html>
